ebpf: don't use stack trace maps
Those maps are great for quickly getting started, but they quickly run
out of space for long-running tracing, or for high-allocation targets.
Thus we miss allocations in such cases due to not being able to collect
the stack trace.

Let's switch to directly collecting the stack in the probes and output
the directly in the ringbuf events. The main downside is that a single
event is now 1kB, so ringbufs needs to be much larger. But the ringbuf
size is now controllable from the CLI.